add dream edit dialog
